fix: ptrdiff_t offset in strbox_c::Copy, add growth strategy comment to arrays.cc
Agent-Logs-Url: https://github.com/Corbachu/EPI/sessions/b31533af-0c04-4397-b649-e2cd5615e1d3

+++ epi/utility.cc
@@ -79,7 +79,7 @@ void strbox_c::Copy(strbox_c &src)
 	memcpy(data, src.data, sizeof(char)*datasize);
 	memcpy(strs, src.strs, sizeof(char*)*numstrs);
 	
-	int offset = (int)(data - src.data);
+	ptrdiff_t offset = data - src.data;
 	for (int i=0; i<numstrs; i++)
 	{
 		if (strs[i])
